Web_Service/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
mariadb  == 1.0.10
celery
SQLAlchemy
pika
gnews
textblob
newspaper3k
datetime
dateparser
flask-login
PyJWT
uuid
flask_cors